+++ HISTORY.rst
@@ -9,8 +9,9 @@ PlatformIO 3.0
 
 * PlatformIO Plus
 
-  + `Unit Testing <http://docs.platformio.org/en/latest/unit_testing.html>`__ for Embedded
-    (`issue #408 <https://github.com/platformio/platformio/issues/408>`_)
+  + Local and Embedded `Unit Testing <http://docs.platformio.org/en/latest/unit_testing.html>`__
+    (`issue #408 <https://github.com/platformio/platformio/issues/408>`_,
+    `issue #519 <https://github.com/platformio/platformio/issues/519>`_)
 
 * Decentralized Development Platforms